fix(Terrain): fixed icons for terrain also added Road-terrain

+++ src/main/java/org/simulation/service/Graphs/DistanceCalculationHeuristic/ChebyshevDistanceCalculator.java
@@ -16,7 +16,7 @@ public class ChebyshevDistanceCalculator implements DistanceCalculator {
     }
 
     @Override
-    public int calculateDistance(Coordinates from, Coordinates to) {
+    public double calculateDistance(Coordinates from, Coordinates to) {
         return Math.max(Math.abs(from.getX() - to.getX()), Math.abs(from.getY() - to.getY()));
     }
 }
