+++ src/main/java/org/simulation/service/Graphs/DistanceCalculationHeuristic/DistanceCalculator.java
@@ -3,5 +3,5 @@ package org.simulation.service.Graphs.DistanceCalculationHeuristic;
 import org.simulation.service.Graphs.Entities.Coordinates;
 
 public interface DistanceCalculator {
-    public int calculateDistance(Coordinates from, Coordinates to);
+    public double calculateDistance(Coordinates from, Coordinates to);
 }
